Predicted and generalized explosive
