+++ code/Weapons/Components/ProjectileComponent.cs
@@ -6,32 +6,35 @@ public partial class ProjectileComponent : WeaponComponent
 	[Prefab]
 	public Model ProjectileModel { get; set; }
 
-	[Prefab]
+	[Prefab, Net]
 	public bool ProjectileShouldBounce { get; set; } = false;
 
-	[Prefab]
+	[Prefab, Net]
 	public bool ProjectileShouldUseTrace { get; set; } = false;
 
-	[Prefab]
+	[Prefab, Net]
 	public int ProjectileMaxBounces { get; set; } = 0;
 
-	[Prefab]
+	[Prefab, Net]
 	public float ProjectileSpeed { get; set; } = 1000.0f;
 
-	[Prefab]
+	[Prefab, Net]
 	public float ProjectileExplosionRadius { get; set; } = 100.0f;
 
-	[Prefab]
+	[Prefab, Net]
 	public float ProjectileExplodeAfter { get; set; } = 4.0f;
 
-	[Prefab]
+	[Prefab, Net]
 	public float ProjectileForceMultiplier { get; set; } = 1.0f;
 
-	[Prefab]
+	[Prefab, Net]
 	public float ProjectileRadius { get; set; } = 1.0f;
 
 	[Prefab, ResourceType( "sound" )]
 	public string ProjectileExplosionSound { get; set; }
+
+	[Prefab, ResourceType( "vpcf" )]
+	public string TrailParticle { get; set; }
 
 	public override bool ShouldStart()
 	{
@@ -48,29 +51,18 @@ public partial class ProjectileComponent : WeaponComponent
 
 	public override void FireInstant()
 	{
-		Log.Info( "Fire Instant" );
+		// Instantly fire using the minimum charge.
+		FireCharged();
 	}
 
 	public override void FireCharged()
 	{
-		Log.Info( "Fire Charged: " + Charge );
-
-		if ( !Game.IsServer )
-			return;
-
 		var position = Weapon.Position.WithY( 0f );
 		var muzzle = Weapon.GetAttachment( "muzzle" );
 		if ( muzzle is not null )
 			position = muzzle.Value.Position.WithY( 0f );
 
-		var projectile = new Projectile()
-			.WithGrub( Grub )
-			.WithModel( ProjectileModel )
-			.WithPosition( position )
-			.WithSpeed( ProjectileSpeed )
-			.WithExplosionSound( ProjectileExplosionSound )
-			.WithExplosionRadius( ProjectileExplosionRadius )
-			.SetCollisionReaction( ProjectileCollisionReaction.Explosive );
+		var projectile = new Projectile();
 
 		if ( ProjectileShouldUseTrace )
 		{
@@ -85,7 +77,7 @@ public partial class ProjectileComponent : WeaponComponent
 		{
 			projectile.SetupPhysicsFromSphere( PhysicsMotionType.Keyframed, position, ProjectileRadius );
 			var desiredPosition = position + (Grub.EyeRotation.Forward.Normal * Grub.Facing * 40f);
-			var tr = Trace.Ray( position, desiredPosition ).Ignore( Weapon.Owner ).Run();
+			var tr = Trace.Ray( position, desiredPosition ).Run();
 			projectile.Position = tr.EndPosition;
 			projectile.Velocity = (Grub.EyeRotation.Forward.Normal * Grub.Facing * Charge * ProjectileSpeed).WithY( 0f );
 		}
@@ -93,11 +85,22 @@ public partial class ProjectileComponent : WeaponComponent
 		if ( ProjectileExplodeAfter > 0 )
 			projectile.ExplodeAfterSeconds( ProjectileExplodeAfter );
 
+		projectile
+			.WithSpeed( ProjectileSpeed )
+			.WithGrub( Grub )
+			.WithModel( ProjectileModel )
+			.WithTrailParticle( TrailParticle )
+			.WithExplosionSound( ProjectileExplosionSound )
+			.WithExplosionRadius( ProjectileExplosionRadius )
+			.SetCollisionReaction( ExplosiveReaction.Explosion );
+
 		projectile.Finish();
 		Grub.SetAnimParameter( "fire", true );
 
 		IsFiring = false;
-		Charge = 0;
+		Charge = MinCharge;
+
+		Grub.CreatedExplosives.Add( projectile );
 
 		FireFinished();
 	}
